requirements.txt
accelerate==1.6.0
aiodns==3.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.3.0
boto3==1.38.13
botocore==1.38.13
Brotli==1.1.0
cached_path==1.7.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
einx==0.3.0
ema-pytorch==0.7.7
email_validator==2.2.0
encodec==0.1.1
fastapi==0.115.12
fastapi-cli==0.0.7
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.0
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.0rc1
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.1
hydra-core==1.3.2
idna==3.10
inquirerpy==0.3.4
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
omegaconf==2.3.0
orjson==3.10.18
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
pfzy==0.3.4
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.8.0
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
pypinyin==0.54.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.5
rsa==4.9.1
ruff==0.11.9
runpod==1.7.9
s3transfer==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentry-sdk==2.27.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
torchdiffeq==0.2.5
tqdm==4.67.1
tqdm-loggable==0.2
transformers==4.51.3
transformers-stream-generator==0.0.5
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vocos==0.1.0
wandb==0.19.11
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
x-transformers==2.3.1
xxhash==3.5.0
yarl==1.20.0

#Video creation
ffmpeg-python==0.2.0